feat: compile warning level, application singleton
compile warning level 3 -> 4 Client, Engine Project both
application structure -> singleton

+++ Project/Engine/Application.cpp
@@ -1,12 +1,13 @@
 #include "Application.h"
+#include "GraphicDeviceDx11.h"
 
 namespace engine
 {
     Application::Application()
         : mGraphicDevice(nullptr)
         , mHwnd(nullptr)
-        , mWidth(-1)
-        , mHeight(-1)
+        , mWidth(static_cast<UINT>(-1))
+        , mHeight(static_cast<UINT>(-1))
         , mbInitialize(false)
     {
     }
@@ -49,13 +50,11 @@ namespace engine
 
             mGraphicDevice = std::make_unique<graphics::GraphicDeviceDX11>();
         }
-
+        
         RECT wdScreen = { 0, 0, static_cast<LONG>(width), static_cast<LONG>(height) };
         AdjustWindowRect(&wdScreen, WS_OVERLAPPEDWINDOW, FALSE);
         SetWindowPos(mHwnd, nullptr /*HWND_TOP*/, 0, 0, wdScreen.right - wdScreen.left, wdScreen.bottom - wdScreen.top, 0/*SWP_NOMOVE | SWP_NOZORDER*/);
         ShowWindow(mHwnd, true);
         UpdateWindow(mHwnd);
-    }
-
-  
+    }    						
 }
